requirements.txt
absl-py==1.4.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
astunparse==1.6.3
beautifulsoup4==4.12.2
blinker==1.6.2
cachelib==0.9.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.4
cmake==3.26.4
colorama==0.4.6
contourpy==1.1.0
coolname==2.2.0
cycler==0.11.0
decorator==5.1.1
deepface==0.0.79
dlib
dnspython==2.3.0
email-validator==2.0.0.post2
filelock==3.12.2
fire==0.5.0
Flask==2.3.2
Flask-Cors==4.0.0
Flask-Mail==0.9.1
Flask-MySQLdb==1.0.1
flask-session2==1.3.1
Flask-WTF==1.1.1
flatbuffers==23.5.26
fonttools==4.41.0
gast==0.4.0
gdown==4.7.1
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.0
gunicorn==20.1.0
h5py==3.9.0
idna==3.4
infinity==1.5
intervals==0.9.2
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.1
keras==2.13.1
kiwisolver==1.4.4
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.2
mtcnn==0.1.1
mysqlclient==2.2.0
nltk==3.8.1
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.8.0.74
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
pipenv==2023.7.11
platformdirs==3.9.1
proctoring==0.0.1
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7.1
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
retina-face==0.0.13
rsa==4.9
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
stripe==5.5.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
termcolor==2.3.0
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.16
validators==0.20.0
virtualenv==20.24.1
virtualenv-clone==0.5.7
Werkzeug==2.3.6
wrapt==1.15.0
WTForms==3.0.1
WTForms-Components==0.10.5